+++ src/parser/connection.rs
@@ -191,6 +191,7 @@ pub(crate) fn connection_def(input: Input<'_>) -> IResult<Input<'_>, Node<Connec
     parse_connection_def(
         input,
         DefinitionPrefixOptions::new(b"connection")
+            .variation_allowed()
             .with_derivation_role()
             .individual_allowed()
             .with_captured_visibility()
@@ -215,6 +216,7 @@ pub(crate) fn connection_def_required(input: Input<'_>) -> IResult<Input<'_>, No
     parse_connection_def(
         input,
         DefinitionPrefixOptions::new(b"connection")
+            .variation_allowed()
             .def_required()
             .individual_allowed()
             .with_captured_visibility()
@@ -235,9 +237,7 @@ fn parse_connection_def(
             start,
             input,
             ConnectionDef {
-                definition_prefix: prefix
-                    .is_abstract
-                    .then_some(crate::ast::DefinitionPrefix::Abstract),
+                definition_prefix: prefix.basic_prefix,
                 is_individual: prefix.is_individual,
                 derivation_role: prefix.derivation_role,
                 identification: prefix.identification,