+++ src/parser/flow.rs
@@ -27,6 +27,7 @@ pub(crate) fn flow_def(input: Input<'_>) -> IResult<Input<'_>, Node<FlowDef>> {
     let (input, prefix) = parse_definition_prefix(
         input,
         DefinitionPrefixOptions::new(b"flow")
+            .variation_allowed()
             .def_required()
             .individual_allowed()
             .with_captured_visibility(),
@@ -38,9 +39,7 @@ pub(crate) fn flow_def(input: Input<'_>) -> IResult<Input<'_>, Node<FlowDef>> {
             start,
             input,
             FlowDef {
-                definition_prefix: prefix
-                    .is_abstract
-                    .then_some(crate::ast::DefinitionPrefix::Abstract),
+                definition_prefix: prefix.basic_prefix,
                 is_individual: prefix.is_individual,
                 identification: prefix.identification,
                 specializes: prefix.specializes,