+++ src/parser/part/def.rs
@@ -7,15 +7,8 @@ pub(crate) fn part_def(input: Input<'_>) -> IResult<Input<'_>, Node<PartDef>> {
     let start = input;
     let (input, _) = ws_and_comments(input)?;
     let (input, (visibility_span, visibility)) = crate::parser::lex::visibility_prefix(input)?;
-    let (input, definition_prefix) = opt(alt((
-        map(preceded(tag(&b"abstract"[..]), ws1), |_| {
-            DefinitionPrefix::Abstract
-        }),
-        map(preceded(tag(&b"variation"[..]), ws1), |_| {
-            DefinitionPrefix::Variation
-        }),
-    )))
-    .parse(input)?;
+    let (input, definition_prefix) =
+        crate::parser::definition_prefix::parse_basic_definition_prefix(input, true)?;
     let (input, is_individual) = opt(preceded(tag(&b"individual"[..]), ws1))
         .parse(input)
         .map(|(i, o)| (i, o.is_some()))?;